Update build.gradle to change targetSDK to 30 instead of 34 - dexError
Task :app:dexBuilderDebug
AGPBI: {"kind":"warning","text":"An API level of 34 is not supported by this compiler. Please use an API level of 30 or earlier","sources":[{}],"tool":"D8"}
AGPBI: {"kind":"warning","text":"An API level of 34 is not supported by this compiler. Please use an API level of 30 or earlier","sources":[{}],"tool":"D8"}
AGPBI: {"kind":"warning","text":"An API level of 34 is not supported by this compiler. Please use an API level of 30 or earlier","sources":[{}],"tool":"D8"}
AGPBI: {"kind":"warning","text":"An API level of 34 is not supported by this compiler. Please use an API level of 30 or earlier","sources":[{}],"tool":"D8"}
AGPBI: {"kind":"warning","text":"An API level of 34 is not supported by this compiler. Please use an API level of 30 or earlier","sources":[{}],"tool":"D8"}

An API level of 34 is not supported by this compiler. Please use an API level of 30 or earlier