Update build.gradle
set the SDK back to 34. complile error. the following error in previous commit was because of old AGP

> Task :app:mergeExtDexDebug
AGPBI: {"kind":"warning","text":"An API level of 34 is not supported by this compiler. Please use an API level of 30 or earlier","sources":[{}],"tool":"D8"}
An API level of 34 is not supported by this compiler. Please use an API level of 30 or earlier